@@ -1,6 +1,6 @@
 {
   "name": "@hoprnet/hopr-testing",
-  "version": "1.68.0-next.18",
+  "version": "1.69.0-next.10",
   "description": "",
   "repository": "https://github.com/hoprnet/hoprnet.git",
   "homepage": "https://hoprnet.org",
@@ -9,13 +9,11 @@
   "types": "lib/index.d.ts",
   "files": [
     "lib/",
-    "!lib/*.spec**",
     "README.md",
-    "package.json",
-    "yarn.lock"
+    "package.json"
   ],
   "engines": {
-    "node": "12"
+    "node": "14"
   },
   "scripts": {
     "clean": "rimraf ./lib",
@@ -23,12 +21,12 @@
     "prepublishOnly": "yarn build"
   },
   "dependencies": {
-    "@hoprnet/hopr-demo-seeds": "1.68.0-next.18",
+    "@hoprnet/hopr-demo-seeds": "1.69.0-next.10",
     "debug": "^4.3.1",
     "ganache-core": "2.13.2"
   },
   "devDependencies": {
-    "@types/node": "^12",
+    "@types/node": "14",
     "rimraf": "^3.0.2",
     "typescript": "^4.1"
   }
