@@ -5,7 +5,7 @@ import type { Connection } from 'libp2p'
 
 const MPLEX = require('libp2p-mplex')
 const KadDHT = require('libp2p-kad-dht')
-const SECIO = require('libp2p-secio')
+import { NOISE } from 'libp2p-noise'
 
 import HoprConnect from '@hoprnet/hopr-connect'
 
@@ -18,7 +18,8 @@ import {
   TICKET_AMOUNT,
   TICKET_WIN_PROB,
   PATH_RANDOMNESS,
-  MIN_NATIVE_BALANCE
+  MIN_NATIVE_BALANCE,
+  FULL_VERSION
 } from './constants'
 
 import { Network } from './network'
@@ -187,16 +188,11 @@ class Hopr<Chain extends HoprCoreConnector> extends EventEmitter {
     const libp2p = await LibP2P.create({
       peerId: id,
       addresses: { listen: addresses },
-      // Disable libp2p-switch protections for the moment
-      switch: {
-        denyTTL: 1,
-        denyAttempts: Infinity
-      },
       // libp2p modules
       modules: {
         transport: [HoprConnect],
         streamMuxer: [MPLEX],
-        connEncryption: [SECIO],
+        connEncryption: [NOISE],
         dht: KadDHT
       },
       config: {
@@ -321,6 +317,12 @@ class Hopr<Chain extends HoprCoreConnector> extends EventEmitter {
     return channels
   }
 
+  /**
+   * Returns the version of hopr-core.
+   */
+  public getVersion() {
+    return FULL_VERSION
+  }
   /**
    * This method starts the node and registers all necessary handlers. It will
    * also open the database and creates one if it doesn't exists.
