@@ -1,12 +1,14 @@
-import assert from 'assert'
+// import { expect } from 'chai'
 import BN from 'bn.js'
-import { stringToU8a, randomInteger } from '@hoprnet/hopr-utils'
+import chai, { expect } from 'chai'
+import chaiAsPromised from 'chai-as-promised'
+import { stringToU8a, u8aToHex } from '@hoprnet/hopr-utils'
 import { Channel, ChannelBalance, ChannelStatus, ChannelState } from './channel'
-import { SignedChannel, Signature, Hash } from '.'
-import * as utils from '../utils'
-import * as testconfigs from '../config.spec'
+// import { Signature, Hash } from '.'
+import Public from './public'
+import SignedChannel from './signedChannel'
 
-const [userA] = testconfigs.DEMO_ACCOUNTS.map((str: string) => stringToU8a(str))
+chai.use(chaiAsPromised)
 
 const generateChannelData = async () => {
   const balance = new ChannelBalance(undefined, {
@@ -21,92 +23,34 @@ const generateChannelData = async () => {
   })
 }
 
-describe('test signedChannel construction', function () {
-  let userAPubKey: Uint8Array
-
-  before(async function () {
-    userAPubKey = await utils.privKeyToPubKey(userA)
-  })
-
+describe('test signedChannel', function () {
   it('should create new signedChannel using struct', async function () {
     const channel = await generateChannelData()
 
-    const signedChannel = new SignedChannel(undefined, {
+    const noCounterparty = new SignedChannel(undefined, {
       channel
     })
-
-    await channel.sign(userA, undefined, {
-      bytes: signedChannel.buffer,
-      offset: signedChannel.signatureOffset
+    const withCounterparty = new SignedChannel(undefined, {
+      channel,
+      counterparty: new Public(stringToU8a('0x03767782fdb4564f0a2dee849d9fc356207dd89f195fcfd69ce0b02c6f03dfda40'))
     })
 
-    assert(await signedChannel.verify(userAPubKey), 'signature must be valid')
-
-    assert(new Hash(await signedChannel.signer).eq(userAPubKey), 'signer incorrect')
-
-    assert(signedChannel.channel.balance.eq(channel.balance), 'wrong balance')
-    assert(new BN(signedChannel.channel._status).eq(new BN(channel._status)), 'wrong status')
-  })
-
-  it('should create new signedChannel using array', async function () {
-    const channel = await generateChannelData()
-
-    const signature = new Signature()
-
-    await channel.sign(userA, undefined, {
-      bytes: signature.buffer,
-      offset: signature.byteOffset
-    })
-
-    const signedChannelA = new SignedChannel(undefined, {
-      signature,
-      channel
-    })
-    const signedChannelB = new SignedChannel({
-      bytes: signedChannelA.buffer,
-      offset: signedChannelA.byteOffset
-    })
-
-    assert(await signedChannelA.verify(userAPubKey), 'signature must be valid')
-    assert(new Hash(await signedChannelA.signer).eq(userAPubKey), 'signer incorrect')
-
-    assert(await signedChannelB.verify(userAPubKey), 'signature must be valid')
-    assert(new Hash(await signedChannelB.signer).eq(userAPubKey), 'signer incorrect')
-
-    assert(signedChannelB.channel.balance.eq(channel.balance), 'wrong balance')
-    assert(new BN(signedChannelB.channel._status).eq(new BN(channel._status)), 'wrong status')
-  })
-
-  it('should create new signedChannel out of continous memory', async function () {
-    const channel = await generateChannelData()
-
-    const signature = new Signature()
-
-    await channel.sign(userA, undefined, {
-      bytes: signature.buffer,
-      offset: signature.byteOffset
-    })
-
-    const offset = randomInteger(1, 31)
-
-    const array = new Uint8Array(SignedChannel.SIZE + offset).fill(0x00)
-
-    const signedChannel = new SignedChannel(
-      {
-        bytes: array.buffer,
-        offset: array.byteOffset + offset
-      },
-      {
-        channel,
-        signature
-      }
+    expect(u8aToHex(await noCounterparty.signer)).to.equal(
+      '0x000000000000000000000000000000000000000000000000000000000000000000'
     )
+    expect(u8aToHex(await withCounterparty.signer)).to.equal(
+      '0x03767782fdb4564f0a2dee849d9fc356207dd89f195fcfd69ce0b02c6f03dfda40'
+    )
+  })
 
-    assert(await signedChannel.verify(userAPubKey), 'signature must be valid')
+  it('should error on verify', async function () {
+    const channel = await generateChannelData()
 
-    assert(new Hash(await signedChannel.signer).eq(userAPubKey), 'signer incorrect')
+    const signedChannel = new SignedChannel(undefined, {
+      channel,
+      counterparty: new Public(stringToU8a('0x03767782fdb4564f0a2dee849d9fc356207dd89f195fcfd69ce0b02c6f03dfda40'))
+    })
 
-    assert(signedChannel.channel.balance.eq(channel.balance), 'wrong balance')
-    assert(new BN(signedChannel.channel._status).eq(new BN(channel._status)), 'wrong status')
+    expect(signedChannel.verify(new Uint8Array())).to.eventually.throw
   })
 })
