@@ -1,16 +1,14 @@
-import { decode } from 'rlp'
 import { createCipheriv, scryptSync, createHmac } from 'crypto'
-import PeerId from 'peer-id'
 
 import {
   KEYPAIR_CIPHER_ALGORITHM,
-  KEYPAIR_SALT_LENGTH,
   KEYPAIR_SCRYPT_PARAMS,
   KEYPAIR_IV_LENGTH,
   KEYPAIR_CIPHER_KEY_LENGTH,
   KEYPAIR_MESSAGE_DIGEST_ALGORITHM
 } from '.'
-import { u8aEquals } from '@hoprnet/hopr-utils'
+
+import { privKeyToPeerId, u8aEquals } from '@hoprnet/hopr-utils'
 
 /**
  * Deserializes a serialized key pair and returns a peerId.
@@ -23,19 +21,25 @@ import { u8aEquals } from '@hoprnet/hopr-utils'
  * @param encryptedSerializedKeyPair the encoded and encrypted key pair
  */
 export async function deserializeKeyPair(encryptedSerializedKeyPair: Uint8Array, password: Uint8Array) {
-  const [salt, mac, encodedCiphertext] = decode(encryptedSerializedKeyPair) as [Buffer, Buffer, Buffer]
-
-  if (salt.length != KEYPAIR_SALT_LENGTH) {
-    throw Error('Invalid salt length.')
-  }
+  const [salt, mac, iv, ciphertext] = [
+    encryptedSerializedKeyPair.subarray(0, 32),
+    encryptedSerializedKeyPair.subarray(32, 64),
+    encryptedSerializedKeyPair.subarray(64, 80),
+    encryptedSerializedKeyPair.subarray(80, 112)
+  ]
 
   const key = scryptSync(password, salt, KEYPAIR_CIPHER_KEY_LENGTH, KEYPAIR_SCRYPT_PARAMS)
 
-  if (!u8aEquals(createHmac(KEYPAIR_MESSAGE_DIGEST_ALGORITHM, key).update(encodedCiphertext).digest(), mac)) {
+  if (
+    !u8aEquals(
+      createHmac(KEYPAIR_MESSAGE_DIGEST_ALGORITHM, key)
+        .update(Uint8Array.from([...iv, ...ciphertext]))
+        .digest(),
+      mac
+    )
+  ) {
     throw Error(`Invalid MAC. Ciphertext might have been corrupted`)
   }
-
-  const [iv, ciphertext] = (decode(encodedCiphertext) as unknown) as [Buffer, Buffer]
 
   if (iv.length != KEYPAIR_IV_LENGTH) {
     throw Error('Invalid IV length.')
@@ -43,5 +47,5 @@ export async function deserializeKeyPair(encryptedSerializedKeyPair: Uint8Array,
 
   let plaintext = createCipheriv(KEYPAIR_CIPHER_ALGORITHM, key, iv).update(ciphertext)
 
-  return await PeerId.createFromProtobuf(plaintext)
+  return await privKeyToPeerId(plaintext)
 }
