@@ -1,7 +1,6 @@
 import { serializeKeyPair } from './serialize'
 import { deserializeKeyPair } from './deserialize'
 
-import { decode, encode } from 'rlp'
 import PeerId from 'peer-id'
 
 import { randomBytes } from 'crypto'
@@ -16,36 +15,45 @@ describe('keypair/index.spec.ts test serialisation and deserialisation of encryp
 
     const peerId = await PeerId.create({ keyType: 'secp256k1' })
 
+    const firstEncoding = serializeKeyPair(peerId, password)
+    const secondEncoding = serializeKeyPair(peerId, password)
+
     assert(
-      !u8aEquals(await serializeKeyPair(peerId, password), await serializeKeyPair(peerId, password)),
+      !u8aEquals(firstEncoding, secondEncoding),
       'Serialization of same peerId should lead to different ciphertexts'
     )
 
-    const serializedKeyPair = await serializeKeyPair(peerId, password)
+    const serializedKeyPair = serializeKeyPair(peerId, password)
     assert(
       u8aEquals((await deserializeKeyPair(serializedKeyPair, password)).marshal(), peerId.marshal()),
       'PeerId must be recoverable from serialized peerId'
     )
 
-    const [salt, mac, encodedCiphertext] = (decode(serializedKeyPair) as unknown) as [Buffer, Buffer, Buffer]
+    const [salt, mac, iv, ciphertext] = [
+      serializedKeyPair.subarray(0, 32),
+      serializedKeyPair.subarray(32, 64),
+      serializedKeyPair.subarray(64, 80),
+      serializedKeyPair.subarray(80, 112)
+    ]
+
     try {
-      const manipulatedSalt = Buffer.from(salt)
-      manipulatedSalt.set(randomBytes(1), randomInteger(0, manipulatedSalt.length))
-      await deserializeKeyPair(encode([manipulatedSalt, mac, encodedCiphertext]), password)
+      const manipulatedSalt = Uint8Array.from(salt)
+      manipulatedSalt.set(randomBytes(1), randomInteger(0, manipulatedSalt.length - 1))
+      await deserializeKeyPair(Uint8Array.from([...manipulatedSalt, ...mac, ...iv, ...ciphertext]), password)
       assert.fail('Shoud fail with manipulated salt')
     } catch {}
 
     try {
-      const manipulatedMac = Buffer.from(salt)
-      manipulatedMac.set(randomBytes(1), randomInteger(0, manipulatedMac.length))
-      await deserializeKeyPair(encode([salt, manipulatedMac, encodedCiphertext]), password)
+      const manipulatedMac = Uint8Array.from(mac)
+      manipulatedMac.set(randomBytes(1), randomInteger(0, manipulatedMac.length - 1))
+      await deserializeKeyPair(Uint8Array.from([...salt, ...manipulatedMac, ...iv, ...ciphertext]), password)
       assert.fail('Shoud fail with manipulated MAC')
     } catch {}
 
     try {
-      const manipulatedCiphertext = Buffer.from(salt)
-      manipulatedCiphertext.set(randomBytes(1), randomInteger(0, manipulatedCiphertext.length))
-      await deserializeKeyPair(encode([salt, mac, manipulatedCiphertext]), password)
+      const manipulatedCiphertext = Uint8Array.from([...iv, ...ciphertext])
+      manipulatedCiphertext.set(randomBytes(1), randomInteger(0, manipulatedCiphertext.length - 1))
+      await deserializeKeyPair(Uint8Array.from([...salt, ...mac, ...manipulatedCiphertext]), password)
       assert.fail('Shoud fail with manipulated ciphertext')
     } catch {}
   })
