@@ -1,4 +1,4 @@
-import Hopr, { FULL_VERSION } from '@hoprnet/hopr-core'
+import type Hopr from '@hoprnet/hopr-core'
 import type HoprCoreConnector from '@hoprnet/hopr-core-connector-interface'
 import { Commands } from './commands'
 import bodyParser from 'body-parser'
@@ -8,7 +8,7 @@ export default function setupAPI(node: Hopr<HoprCoreConnector>, logs: any, optio
   const service = require('restana')()
   service.use(bodyParser.text({ type: '*/*' }))
 
-  service.get('/api/v1/version', (_, res) => res.send(FULL_VERSION))
+  service.get('/api/v1/version', (_, res) => res.send(node.getVersion()))
   service.get('/api/v1/address/eth', async (_, res) => res.send(await node.paymentChannels.hexAccountAddress()))
   service.get('/api/v1/address/hopr', async (_, res) => res.send(await node.getId().toB58String()))
 
