@@ -1,7 +1,11 @@
 import { AbstractCommand } from './abstractCommand'
-import { FULL_VERSION } from '@hoprnet/hopr-core'
+import type HoprCoreConnector from '@hoprnet/hopr-core-connector-interface'
+import type Hopr from '@hoprnet/hopr-core'
 
 export default class Version extends AbstractCommand {
+  constructor(public node: Hopr<HoprCoreConnector>) {
+    super()
+  }
   public name() {
     return 'version'
   }
@@ -11,6 +15,6 @@ export default class Version extends AbstractCommand {
   }
 
   public async execute(): Promise<string> {
-    return FULL_VERSION
+    return this.node.getVersion()
   }
 }
